Added argument `droptables`, fixed `dbconfig` argument

+++ wfs-downloader.py
@@ -1,10 +1,10 @@
 import os
 import json
 import time
-import geojson
 import argparse
-import tempfile
 import traceback
+
+import geojson
 from colorama import init, Fore, Style
 from owslib.wfs import WebFeatureService
 
@@ -45,19 +45,25 @@ def main():
             help="Overwrite if already exists the file or the table (default: %(default)s)",
         )
         parser.add_argument(
+            "--droptables",
+            action="store_true",
+            help="Delete the existing tables instead of only cleaning/adding the values. Only used if `overwrite` is active and store to table is configured (default: %(default)s)",
+        )
+        parser.add_argument(
             "--sleep",
             type=float,
             metavar="Sleep time",
             default=sleep,
             help="Sleep time (in seconds) between each reques to avoid overloading the server (default: %(default)s)",
         )
-  
+
         args = parser.parse_args()
 
         config = args.config
         overwrite = args.overwrite
+        drop_tables = args.droptables
         sleep = args.sleep
-        database = args.database
+        database = args.dbconfig
 
         print("--> PROCESS STARTED <--")
         print("\t")
@@ -74,6 +80,7 @@ def main():
         layers_group = data["group"]
 
         for group in layers_group:
+            print(f"Connecting to the server {group['url']}")
 
             version = group["version"] or "1.1.0"
             wfs = WebFeatureService(url=group["url"], version=version)
@@ -121,22 +128,39 @@ def main():
                             table = target["table"]
                             schema = target["schema"] or "public"
 
+                            if not table:
+                                print(f"--> Table name is not specified, skipping...")
+                                continue
+
                             sql = """
                                 SELECT EXISTS (
                                     SELECT FROM information_schema.tables 
                                     WHERE  table_schema = %s
                                     AND    table_name   = %s
                                 )"""
-                            
-                            db.cur.execute(sql, (schema, table))                         
+
+                            db.cur.execute(sql, (schema, table))
                             table_exists = db.cur.fetchone()[0]
-                            
+
+                            create_table = True
+
                             if table_exists:
                                 if overwrite:
-                                    db.cur.execute(
-                                        f"DROP TABLE IF EXISTS {schema}.{table} CASCADE;"
-                                    )
-                                    print(f"--> Removing old table {schema}.{table}")
+                                    if drop_tables:
+                                        db.cur.execute(
+                                            f"DROP TABLE IF EXISTS {schema}.{table} CASCADE;"
+                                        )
+                                        print(
+                                            f"--> Removing old table {schema}.{table}"
+                                        )
+                                    else:
+                                        create_table = False
+                                        db.cur.execute(
+                                            f"DELETE * FROM {schema}.{table};"
+                                        )
+                                        print(
+                                            f"--> Removing values from {schema}.{table}"
+                                        )
                                 else:
                                     print(
                                         f"--> Skiping existing table {schema}.{table}"
@@ -148,31 +172,31 @@ def main():
                             db.cur.execute(f"CREATE SCHEMA IF NOT EXISTS {schema}")
 
                             data = geojson.loads(geojson_data)
-
                             sample_feature = data["features"][0]
                             properties = sample_feature["properties"]
 
-                            columns = []
-                            for key, value in properties.items():
-                                if isinstance(value, int):
-                                    column_type = "INTEGER"
-                                elif isinstance(value, float):
-                                    column_type = "FLOAT"
-                                elif isinstance(value, bool):
-                                    column_type = "BOOLEAN"
-                                else:
-                                    column_type = "TEXT"
-                                columns.append(f'"{key}" {column_type}')
+                            if create_table:
+                                columns = []
+                                for key, value in properties.items():
+                                    if isinstance(value, int):
+                                        column_type = "INTEGER"
+                                    elif isinstance(value, float):
+                                        column_type = "FLOAT"
+                                    elif isinstance(value, bool):
+                                        column_type = "BOOLEAN"
+                                    else:
+                                        column_type = "TEXT"
+                                    columns.append(f'"{key}" {column_type}')
 
-                            create_table_query = f"""
-                            CREATE TABLE {schema}.{table} (
-                                id SERIAL PRIMARY KEY,
-                                geom GEOMETRY(Geometry, 4326),
-                                {", ".join(columns)}
-                            )
-                            """
-                            db.cur.execute(create_table_query)
-                            print(f"Table created {schema}{table}.")
+                                create_table_query = f"""
+                                CREATE TABLE {schema}.{table} (
+                                    id SERIAL PRIMARY KEY,
+                                    geom GEOMETRY(Geometry, 4326),
+                                    {", ".join(columns)}
+                                )
+                                """
+                                db.cur.execute(create_table_query)
+                                print(f"Table created {schema}{table}.")
 
                             insert_query = f"""
                                 INSERT INTO {schema}.{table} (geom, {", ".join(f'"{key}"' for key in properties.keys())})
@@ -183,7 +207,7 @@ def main():
                                 geom = geojson.dumps(feature["geometry"])
                                 values = tuple(properties.values())
                                 db.cur.execute(insert_query, (geom, *values))
-                            
+
                             print(f"Inserted entries in {schema}{table}.")
                             db.conn.commit()
